# fix dataloader performace decrease & unittest error